Add getter for any content in the CategoryEntriesModel, and isBook checker
There's logic here which is not entirely sensible, and which needs fixing.
It works for now, but it is starting to look like refactoring is needed.
This also goes for other parts of the code, and there is no sensible reason
for the split between category and categoryentries models, now that the
latter is recursive...