Animate the page-changes and thumbnail movement
Why this is not something that can be done automatically i will never
understand. In a toolset made specifically to make animations of your
UI easy and straightforward, having to do this just feels like an
enormous hack. I can see why it is perhaps not default behaviour, but
really, why not at least make it an option?