Use a FileDialog to allow adding new folders to the list
PeruseConfig will take any file:// url and add that, ensuring
that it only adds things not already in the list, and removing
duplicates and things covered by the new addition.

This also introduces a reporting feature to peruseconfig, which
is currently unused, but which i shall be connecting up later.