Switch fullscreen when showing/hiding controls
The idea behind this is that when you hide the controls on the
application window, what you are in fact asking for is for
everything to stop bothering you. In short, immerse yourself
in things and just experience the content, nothing else. So,
hide chrome, full screen, and just read.