RightToLeft mode for comic reading
This is really just layout, but, call it Manga Mode, it makes
reading things laid out right to left style feel much more natural.
Thank you to bodertz for the suggestion!

Caveat: This should be stored per-comic, but isn't currently. This
will change when the more complete metadata support is done.

BUG:364682