Add documentation to the QtQuick and ContentList plugins.
This adds some documentation strings to the header files in the ContentList folder and the QtQuick folder.
TODO entries were put wherever I just had no clue what a function or the like did, meaning they are actually
a little more in need of having docs than any of the other things.

Differential Revision: https://phabricator.kde.org/D15264