Make retrieving title, keywords and annotation a bit more robust.
This will first check if the requested language is available, and if
not set to default.

If the default is empty, it will try to get the entry of the first language
in the list of languages.

If that is empty too, it will try to get the first entry it can find.

We could in the future extend this by trying to implement fallback
languages(important for the variants of Portguese and Chinese), but
right now this will try all possible ways to return some data.