Add adding/removing jumps, text-areas and frames.
Clicking once on the page sheet will give a selection rectangle.
Clicking again somewhere will confirm the rectangle and ask whether you wanna
make a frame/textarea/jump.

It then finalises the selection, and you can further edit the item in the page
info editor.

They save properly too. It's not the most self-evident workflow (not mobile friendly much),
but it is pretty passable to use! :)