Make swaps work for page areas by using stringlists instead of ints.
When just counting and checking the changes for the total amount of a given
area (frame, text area, jump), QML repeaters and listviews will not change
when the areas on a page are swapped. After all, the count has not changed.

So instead we're now keeping track of stringlists consisting of the points.
This is both a little hacky, but it also makes sense: the areas are a bit more
than a number, so keeping stringlists does make a little sense.

We wanted to avoid using QQMLListProperty, as that would require making the acbf
library into a QML plugin.