Frame based zoom in the image browser.
If there's ACBF frama data available, the Image Browser will now keep track of the location of
said frames, and when those locations are double clicked, zoom in till the frame fits the viewport.

There's stuff in place to switch to the next frame, but not really hooked up to anything yet.

Differential Revision: https://phabricator.kde.org/D15661