Support multiple authors/ multiple series/description and ratings/tags/comments.
This patch adds support for multiple authors and series. Description it gets from the acbf annotation.

Tags/Comments/Ratings all come from KFileMetaData::UserData, being the same items you see when looking
at an file in dolphin. You can now set these from Peruse, which will be useful when dealing with indie
stuff where there's lots of single books that you'd want to annotate.

Because of the way how the database works, we get the ratings/tags/comments when checking a booktile

The database still doesn't update, but we do update the relevant entries in the booklistmodel.

Differential Revision: https://phabricator.kde.org/D15713