BUG:398260 avoid removed entries in the cache
This removes entries from the cache when removing books within peruse,
but also removes entries when QFileInfo::exists() cannot verify they exist(when removed
from outside).

The latter might be a little slower, but not as noticable as the slowness of building the
cache for the first time. If it does become a problem this will be the first place to start
optimising.

Differential Revision: https://phabricator.kde.org/D15748