Migrate logging to categorized one
Add autogenerated logging categories for the main areas of the peruse
sources, and switch the various qDebug/qWarning to use them.  This way,
peruse is silent by default, and it is possible to enable the debug
changing the Qt logging rules (e.g. via envvar).

Remove all the QDebug includes, since they are no more needed.