Fix some sporadic crashes in the previewing code
This has been a very long standing issue, and essentially a release
blocker. Finally worked out that an underdocumented class named
QQuickAsyncImageProvider is supposed to be used for things like this.
So, use that, and the crashes go away (and the UI is smoother in
addition, which is always nice).

BUG:398204