Handle layers for which is the current page
This works around an issue in Kirigami, where Page.iscurrentPage
does not contain support for layers, and returns true even if
the current layer is not the one that page is found on. This
may be by design, but it is awkward and we need to ensure that
it is somehow explicit, but this workaround functions for now.