Make a bit of useful noise when a page's image doesn't load
This was reported by Luca, who discovered that without the Qt5
ImageFormats package installed, Peruse would silently fail to display
pages in books which use e.g. webp as the page image format. This
affects, for example, the Pepper & Carrot comics on the KDE Store, and
given they're our go-to for demoing, seems like something to catch.