Add fileEntryReferenced function to archive book (and add webp to images)
This function allows us to check whether or not a file entry in the
archive is referenced somewhere in the ACBF data (which is useful for
linting purposes and whatnot, and just generally handy). It works at two
levels, and reports both whether there is a reference by its full path
name, or in a stylesheet by just filename. It is a bit fuzzy, but
important. Testing done on the Pepper and Carrot book 1 from the Store.