Makes Reader support visualizing Jump objects
This commit extends Reader to support visualizing Jumps on Pages/Frames, by
adding a new component called JumpHandler. JumpHandler would take in the actual
Jump object along with the offset, and zooming parameters to draw itself as an
animated circular marker that gets highlighted. The highlighting is configured
as an application settings that has been added in PeruseConfig called "animate
jump areas". In order to integrate the new component, we had to modify the ImageBrowser
by adding a Repeater instance that holds enough JumpHandlers for each page