Resets the actual fields instead of AddPageArea properties
Before this commit, BookPage would reset every property inside
AddPageArea resulting in breaking the bindings of these properties and
in them having default values. So, we end up with a new page area object
but with its default values submitted to our model. This commit fixes
the mentioned issue by actually resetting the inner fields of the
components of AddPageArea so we don't break the originally created
bindings