Implement the perfectionist line fitting
Also adds in functionality so we push down the line until it fits the
first word (as decided by average character width for the font
multiplied by the text length plus one of the current line). This
ensures we are able to more gracefully fit at least one word on a line,
which yields larger font sizes, and better general layouting.