Extends Peruse to support jumps interactions using keyboard
This commit extends the ViewerBase to have support for keyboard interactions by adding a new list property representing the new actions to be added by the ViewerBase implementors. The cbr component now adds new two actions for handling hitting the "Tab", and the "Return" keys on the keyboard, to switch and activate the jumps accordingly. We also refactored the ImageBrowser a bit by which we initialize each frame on each "frame change" event that occurs, in which we only store only the jumps/objects that should be visible/contained by this frame and initialize the currentJumpIndex which helps us in switching/activating jumps