Extends Creator with a new editing form
This creates a new form called EditPageArea to allow for a convenient
and easier way to edit/modify particular objects on pages upon clicking.
Heavily inspired by the original AddPageArea, it changes its layout
according to the given object's type. This also modifies the original
AddPageArea, in which it now only have three buttons to choose the type
of the object from. So, the approach for adding a new object now, is
creating a new page area, followed by sending the created object to the
new EditPageArea form to modify its initial values.